Extract holder mediation methods to separate file
