+++ aries_vcx/tests/utils/scenarios.rs
@@ -35,9 +35,9 @@ pub mod test_utils {
     use aries_vcx::common::proofs::proof_request::PresentationRequestData;
     use aries_vcx::common::proofs::proof_request_internal::AttrInfo;
     use aries_vcx::handlers::connection::mediated_connection::{ConnectionState, MediatedConnection};
-    use aries_vcx::handlers::issuance::holder::test_utils::get_credential_offer_messages;
     use aries_vcx::handlers::issuance::holder::Holder;
     use aries_vcx::handlers::issuance::issuer::Issuer;
+    use aries_vcx::handlers::issuance::mediated_holder::get_credential_offer_messages;
     use aries_vcx::handlers::issuance::mediated_issuer::{get_credential_proposal_messages, issuer_update_with_mediator};
     use aries_vcx::handlers::proof_presentation::prover::test_utils::get_proof_request_messages;
     use aries_vcx::handlers::proof_presentation::prover::Prover;
